back/requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=dev.txt common.in dev.in
#
aiohttp==3.9.1
    # via black
aiosignal==1.3.1
    # via aiohttp
argon2-cffi==23.1.0
    # via -r common.in
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asgiref==3.7.2
    # via
    #   django
    #   django-cors-headers
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
black==23.12.0
    # via -r dev.in
build==1.0.3
    # via pip-tools
cffi==1.16.0
    # via argon2-cffi-bindings
click==8.1.7
    # via
    #   black
    #   pip-tools
coverage==7.3.3
    # via -r dev.in
django==4.2
    # via
    #   -r common.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==4.3.1
    # via -r common.in
djangorestframework==3.14.0
    # via -r common.in
flake8==6.1.0
    # via -r dev.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
idna==3.6
    # via yarl
isort==5.13.2
    # via -r dev.in
mccabe==0.7.0
    # via flake8
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via black
packaging==23.2
    # via
    #   black
    #   build
pathspec==0.12.1
    # via black
pip-tools==7.3.0
    # via -r dev.in
platformdirs==4.1.0
    # via black
psycopg2-binary==2.9.9
    # via -r common.in
pycodestyle==2.11.1
    # via flake8
pycparser==2.22
    # via cffi
pyflakes==3.1.0
    # via flake8
pyproject-hooks==1.0.0
    # via build
python-dotenv==1.0.0
    # via -r common.in
python-magic==0.4.27
    # via -r common.in
pytz==2023.3.post1
    # via djangorestframework
sqlparse==0.4.4
    # via django
tomli==2.0.1
    # via
    #   black
    #   build
    #   pip-tools
    #   pyproject-hooks
typing-extensions==4.9.0
    # via
    #   asgiref
    #   black
wheel==0.42.0
    # via pip-tools
yarl==1.9.4
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
